Moved NumberAbreviation function to utils
